+++ static/onlyoffice-docs-confluence-cloud-custom-ui/src/components/ContentTree/utils/iconUtils.tsx
